pkg/pyproject.toml
[project]
name = "db_oracle_get_mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.11.0",
    "oracledb>=3.2.0",
    "python-dotenv>=1.1.1",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[project.scripts]
db_oracle_get_mcp = "pkg.db_oracle_get_mcp:main"
dev = "pkg.db_oracle_get_mcp_sse:dev"